requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
argcomplete==1.10.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.8.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
compressed-rtf==1.0.6
dataclasses-json==0.5.14
Deprecated==1.2.14
diskcache==5.6.3
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.3
extract-msg==0.28.7
fastapi==0.104.1
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
IMAPClient==2.1.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.327
langsmith==0.0.56
llama-index==0.8.57
llama_cpp_python==0.2.12
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
olefile==0.46
openai==0.28.1
packaging==23.2
pandas==2.1.2
pdfminer.six==20191110
Pillow==9.5.0
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.0.1
python-multipart==0.0.6
python-pptx==0.6.22
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.22
starlette==0.27.0
tenacity==8.2.3
textract==1.6.5
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.23.2
wrapt==1.15.0
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.2
